src/main/java/com/ec/sticket/controllers/MainController.java
package com.ec.sticket.controllers;

import com.ec.sticket.models.User;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/main")
public class MainController {

    @GetMapping("/user")
    public User getUser(){
        return new User();
    }


}
